[#33331] explicit integer in some header files

+++ bibfor/include/asterfort/amumpm.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2021 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -24,10 +24,10 @@ interface
                       klag2, type, lmd, epsmat, ktypr,&
                       lpreco, lmhpc, lbloc)
         aster_logical :: ldist
-        integer :: kxmps
+        integer(kind=8) :: kxmps
         character(len=24) :: kmonit(12)
         character(len=14) :: impr
-        integer :: ifmump
+        integer(kind=8) :: ifmump
         character(len=5) :: klag2
         character(len=1) :: type
         aster_logical :: lmd
