+++ bibfor/include/asterfort/assert.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2022 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -57,6 +57,6 @@ interface
         aster_logical :: cond
         character(len=*) :: str_cond
         character(len=*) :: fname
-        integer :: line
+        integer(kind=8) :: line
     end subroutine assert
 end interface
